+++ lib/LibCrafts-1.0/LibCrafts-1.0.lua
@@ -11,7 +11,7 @@
 local LibStub = getglobal("LibStub")
 assert(LibStub ~= nil)
 
-local untyped_lib, _ = LibStub:NewLibrary("LibCrafts-1.0", 5)
+local untyped_lib, _ = LibStub:NewLibrary("LibCrafts-1.0", 7)
 if not untyped_lib then return end
 
 ---@class LibCrafts
@@ -20,12 +20,14 @@ if not untyped_lib then return end
 ---@field modules_by_name table<string, LcModule>
 ---@field spell_id_to_craft table<number, LcCraft>
 ---@field reagent_id_to_spell_ids_set table<number, table<number, boolean>>
+---@field recipe_id_to_spell_ids_set table<number, table<number, boolean>>
 ---@field localized_profession_name_to_spell_ids_set table<string, table<number, boolean>>
 
 local lib = --[[---@type LibCrafts]] untyped_lib
 lib.modules_by_name = lib.modules_by_name or {}
 lib.spell_id_to_craft = lib.spell_id_to_craft or {}
 lib.reagent_id_to_spell_ids_set = lib.reagent_id_to_spell_ids_set or {}
+lib.recipe_id_to_spell_ids_set = lib.recipe_id_to_spell_ids_set or {}
 lib.localized_profession_name_to_spell_ids_set = lib.localized_profession_name_to_spell_ids_set or {}
 
 ---@type table[]
@@ -33,6 +35,7 @@ local data_tables = {
     lib.modules_by_name,
     lib.spell_id_to_craft,
     lib.reagent_id_to_spell_ids_set,
+    lib.recipe_id_to_spell_ids_set,
     lib.localized_profession_name_to_spell_ids_set,
 }
 
@@ -48,12 +51,14 @@ if not all_data_tables_filled then
     lib.modules_by_name = {}
     lib.spell_id_to_craft = {}
     lib.reagent_id_to_spell_ids_set = {}
+    lib.recipe_id_to_spell_ids_set = {}
     lib.localized_profession_name_to_spell_ids_set = {}
 end
 
 lib.env = {
     is_debug = false,
     is_turtle_wow = getglobal("LFT") ~= nil,
+    is_super_wow_loaded = getglobal("SpellInfo") ~= nil,
 }
 
 lib.constants = {
@@ -86,6 +91,7 @@ lib.constants = {
 ---@shape LcEnvironment
 ---@field is_debug boolean
 ---@field is_turtle_wow boolean
+---@field is_super_wow_loaded boolean
 
 ---@shape LcConstants
 ---@field qualities LcItemQualities
@@ -139,6 +145,7 @@ lib.constants = {
 ---@field sources LcSpellSource[]
 ---@field recipes LcRecipe[]
 ---@field reagent_id_to_count table<number, number>
+---@field was_enriched boolean
 local Craft = {}
 
 ---@param value string
@@ -160,6 +167,22 @@ local function translate_from_en_to_game_locale(value, locale_module_name)
     return --[[---@type string]] localized_value
 end
 
+---@param craft LcCraft
+---@return LcCraft
+local function enrich(craft)
+    if not lib.env.is_super_wow_loaded or craft.was_enriched then
+        return craft
+    end
+
+    local spell_name, _, _, _, _ = SpellInfo(craft.spell_id)
+    if spell_name ~= nil then
+        craft.localized_spell_name = spell_name
+        craft.was_enriched = true
+    end
+
+    return craft
+end
+
 ---
 --- Returns a list of crafts by reagent item id.
 ---
@@ -168,7 +191,20 @@ end
 function lib:GetCraftsByReagentId(item_id)
     local crafts = {}
     for spell_id, _ in pairs(self.reagent_id_to_spell_ids_set[item_id] or {}) do
-        tinsert(crafts, self.spell_id_to_craft[spell_id])
+        tinsert(crafts, enrich(self.spell_id_to_craft[spell_id]))
+    end
+    return crafts
+end
+
+---
+--- Returns a list of crafts by recipe item id.
+---
+---@param item_id number
+---@return LcCraft[]
+function lib:GetCraftsByRecipeId(item_id)
+    local crafts = {}
+    for spell_id, _ in pairs(self.recipe_id_to_spell_ids_set[item_id] or {}) do
+        tinsert(crafts, enrich(self.spell_id_to_craft[spell_id]))
     end
     return crafts
 end
@@ -183,7 +219,7 @@ function lib:GetCraftsByProfession(profession)
     local spell_ids_set = d[profession] or d[translate_from_en_to_game_locale(profession, "Locales-Professions")] or {}
     local crafts = {}
     for spell_id, _ in pairs(spell_ids_set) do
-        tinsert(crafts, self.spell_id_to_craft[spell_id])
+        tinsert(crafts, enrich(self.spell_id_to_craft[spell_id]))
     end
     return crafts
 end
@@ -395,11 +431,21 @@ function Craft:Save()
         assert(next(self.reagent_id_to_count) ~= nil)
     end
 
+    self.was_enriched = false
+    if lib.env.is_super_wow_loaded then
+        enrich(self)
+    end
+
     local old_craft = lib.spell_id_to_craft[self.spell_id]
     if old_craft ~= nil then
         for reagent_id, _ in pairs(old_craft.reagent_id_to_count) do
             lib.reagent_id_to_spell_ids_set[reagent_id][self.spell_id] = nil
         end
+
+        for _, recipe in ipairs(old_craft.recipes) do
+            lib.recipe_id_to_spell_ids_set[recipe.id][self.spell_id] = nil
+        end
+
         lib.localized_profession_name_to_spell_ids_set[old_craft.localized_profession_name][self.spell_id] = nil
     end
 
@@ -408,6 +454,11 @@ function Craft:Save()
     for reagent_id, _ in pairs(self.reagent_id_to_count) do
         get_or_create_set(lib.reagent_id_to_spell_ids_set, reagent_id)[self.spell_id] = true
     end
+
+    for _, recipe in ipairs(self.recipes) do
+        get_or_create_set(lib.recipe_id_to_spell_ids_set, recipe.id)[self.spell_id] = true
+    end
+
     get_or_create_set(lib.localized_profession_name_to_spell_ids_set, self.localized_profession_name)[self.spell_id] = true
 end