+++ lib/LibCrafts-1.0/Professions/Cooking.lua
@@ -1,5 +1,5 @@
 local lib = LibCraftsGetLibrary()
-local name, version = "Professions-Cooking", 2
+local name, version = "Professions-Cooking", 3
 local module = --[[---@type LcProfessionModule]] lib:RegisterProfessionModule(name, version, "Cooking")
 if not module then return end
 
@@ -592,6 +592,15 @@ if lib.env.is_turtle_wow then
         :AddReagent(12037, 2) -- Mystery Meat
         :AddReagent(12202, 1) -- Tiger Meat
         :Save()
+
+    module:NewCraft(49551, "Empowering Herbal Salad", 300, {})
+          :SetResult(83309)
+          :AddRecipe(92045, Quality.Rare, {RecipeSource.Quest})
+          :AddReagent(6370, 1) -- Blackmouth Oil
+          :AddReagent(8838, 1) -- Sungrass
+          :AddReagent(22529, 1) -- Savage Frond
+          :AddReagent(51714, 3) -- Sweet Mountain Berry
+          :Save()
 
     module:NewCraft(57047, "Danonzo's Tel'Abim Surprise", 300, {})
         :SetResult(60976)