+++ lib/LibCraftingProfessions-1.0/CHANGES.md
@@ -1,3 +1,6 @@
+### Revision 11 (2025/06/01)
+- Add profession icon textures
+
 ### Revision 10 (2025/05/30)
 - Add profession frame SHOW/CLOSE events
 - Support all modern major 1.12 profession frame replacement addons with all their quirks and little details ([AdvancedTradeSkillWindow](https://github.com/laytya/AdvancedTradeSkillWindow-vanilla), [AdvancedTradeSkillWindow2](https://github.com/Shellyoung/AdvancedTradeSkillWindow2), [Artisan](https://github.com/Otari98/Artisan))